REF: split crop_and_bin into two separate widgets
